move SafeMap to common folder
